requirements.txt
streamlit
langchain
langchain-community
langchain-ollama
langchain-text-splitters
faiss-cpu
pypdf
PyPDF2
python-docx
huggingface-hub
sentence-transformers